Allow multiple processes accessing the same Redis database by using redis pipeline. By default all pipelined commands are executed atomically
