+++ docs/source/reference/storages.rst
@@ -4,4 +4,4 @@ Storages
 ========
 
 .. autoclass:: RDBStorage
-.. autoclass:: optuna.storages.redis_storage.RedisStorage
+.. autoclass:: RedisStorage
