+++ optuna/storages/__init__.py
@@ -1,6 +1,6 @@
 from optuna.storages.base import BaseStorage  # NOQA
 from optuna.storages.in_memory import InMemoryStorage  # NOQA
-from optuna.storages.redis_storage import RedisStorage  # NOQA
+from optuna.storages.redis import RedisStorage  # NOQA
 from optuna.storages.rdb.storage import RDBStorage  # NOQA
 
 from typing import Union  # NOQA
@@ -12,8 +12,8 @@ def get_storage(storage):
     if storage is None:
         return InMemoryStorage()
     if isinstance(storage, str):
-        if storage.startswith('redis'):
-            return RedisStorage(url=storage)
+        if storage.startswith("redis"):
+            return RedisStorage(storage)
         else:
             return RDBStorage(storage)
     else:
